This contains the code for the Split the Bill App
Split The Bill is a java based android program created through the use of the program Android Studio. In this app there are two different pages for the user to experience. On the first page there is a logo for the app and a button allow the user to transition to the second page. The second page allows the user to enter a bill amount, the number of party members and allows a user to select an experience type. When the user presses the calculation button, the app will calculated the total bill by adding a base tip amount to the user’s bill and divides the bill by the total number of party members. After the calculation, the application will tell the user below the calculation button what the tip, total bill and individual cost are.